LayoutTests/platform/ios-sim/accessibility/disabled-states-expected.txt

button 1
button 2
button 3
 
This test thats disabled and aria-disabled change the traits returned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS btn1.iphoneTraits != btn2.iphoneTraits is true
PASS btn2.iphoneTraits == btn3.iphoneTraits is true
PASS select1.iphoneTraits != select2.iphoneTraits is true
PASS successfullyParsed is true

TEST COMPLETE
